requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.2.0
crispy-bootstrap5==0.7
Django==4.2.5
django-auto-logout==0.5.1
django-axes==6.1.1
django-crispy-forms==2.0
django-formtools==2.4.1
django-otp==1.2.2
django-paypal==2.0
django-phonenumber-field==6.4.0
django-recaptcha==3.0.0
django-two-factor-auth==1.15.4
idna==3.4
phonenumbers==8.13.20
Pillow==10.0.0
pypng==0.20220715.0
python-dotenv==1.0.0
pytz==2023.3.post1
qrcode==7.4.2
requests==2.31.0
sqlparse==0.4.4
stripe==6.5.0
typing_extensions==4.7.1
urllib3==2.0.4
